refactor(objc): split rules and origin classification
