--- Sources/ObjCSyntaxInk/ObjCSemanticTokenProvider.swift
@@ -1,11 +1,0 @@
-import Foundation
-
-enum ObjCSemanticTokenProvider {
-    static func semanticTokens(for source: String) -> [ObjCSemanticToken]? {
-#if os(macOS)
-        SourceKitLSPClient.shared.semanticTokens(for: source, kind: .infer(from: source))
-#else
-        nil
-#endif
-    }
-}
